add autocorrelation functions to the Cells files in the DB and put them in the cells table
by default, calculated for the first 0.5s of fixation. violation trials and any trials where clicks are present in that interval are excluded. because of some settings files that had a small number of trials with clicks starting immediately upon fixation (A241 and A243) this step needed to be taken.